site wide: New Apache Software Foundation (ASF) logo
The Apache Software Foundation (ASF) has introduced a new logo for the
foundation, replacing the old logo. Read the announcement at [1].

Updating the Apache NuttX website to display the new foundation logo.

Prior to this change, this website used a PNG file to display the old
logo with a width of 439 pixels. To retain the same format and size,
the new logo was converted from the official SVG (see [2]) to PNG
format (retaining the transparent background) using Inkscape 1.2.2 with
the following invocation:

    $ inkscape -w 439 asf_logo_wide.svg -o asf-logo-wide.png

Details of changes:

* assets/themes/apache/img/asf-estd-1999-logo.png: Deleted.

* assets/themes/apache/img/asf-logo-wide.png: New.

* _includes/themes/apache/footer.html:
  Bump copyright year.
  Update trademark notice to match text in the Style Guide (see [3]).

References:

[1] New logo announcement:
    https://news.apache.org/foundation/entry/introducing-the-asfs-new-logo

[2] SVG original:
    https://www.apache.org/foundation/press/kit/asf_logo_wide.svg

[3] Style Guide:
    https://www.apache.org/foundation/press/kit/ApacheFoundation_StyleGuide.pdf